pyproject.toml
[project]
name = "nonebot-plugin-tvseries"
version = "0.0.2.3"
description = ""
authors = [
    {name = "rf_tar_railt", email = "3165388245@qq.com"},
    {name = "kexue", email = "xana278@foxmail.com"},
]
dependencies = [
    "lxml<5.0.0,>=4.6.4",
    "nonebot-plugin-htmlrender",
    "nonebot2>=2.3.0",
    "nonebot-plugin-alconna>=0.45.4",
]
requires-python = "<4.0,>=3.9"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
dev = [
    "black",
    "pillow<10.0.0",
    "nonebot2[httpx]"
]

[tool.pdm.build]
includes = ["src"]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

